move majority of campaign related logic in game_newgame to game_campaign
